server/pyproject.toml
[project]
name = "nebula-core"
version = "0.1.0"
description = "MCP server for Nebula, a privacy-first memory layer with approval workflows for AI agents"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [
  { name = "alxx", email = "contact@alxx.dev" },
]

dependencies = [
  "mcp[cli]>=1.26.0",
  "asyncpg>=0.30.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.12.5",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pyfiglet>=1.0.2",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.34.0",
  "argon2-cffi>=23.1.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["ruff", "black"]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
markers = [
    "unit: no DB needed",
    "integration: requires real PostgreSQL",
    "database: SQL-level tests",
    "e2e: full workflow tests",
    "api: REST API endpoint tests",
]

[dependency-groups]
dev = [
    "black>=24.0.0",
    "ruff>=0.3.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
]
